+++ cmd/neofs-cli/modules/object/head.go
@@ -10,9 +10,11 @@ import (
 	"github.com/nspcc-dev/neofs-node/cmd/neofs-cli/internal/common"
 	"github.com/nspcc-dev/neofs-node/cmd/neofs-cli/internal/commonflags"
 	"github.com/nspcc-dev/neofs-node/cmd/neofs-cli/internal/key"
+	"github.com/nspcc-dev/neofs-sdk-go/client"
 	cid "github.com/nspcc-dev/neofs-sdk-go/container/id"
 	"github.com/nspcc-dev/neofs-sdk-go/object"
 	oid "github.com/nspcc-dev/neofs-sdk-go/object/id"
+	"github.com/nspcc-dev/neofs-sdk-go/user"
 	"github.com/spf13/cobra"
 )
 
@@ -37,7 +39,6 @@ func initObjectHeadCmd() {
 	_ = objectHeadCmd.MarkFlagRequired(commonflags.OIDFlag)
 
 	flags.String(fileFlag, "", "File to write header to. Default: stdout.")
-	flags.Bool("main-only", false, "Return only main fields")
 	flags.Bool(commonflags.JSON, false, "Marshal output in JSON")
 	flags.Bool("proto", false, "Marshal output in Protobuf")
 	flags.Bool(rawFlag, false, rawFlagDesc)
@@ -47,11 +48,11 @@ func getObjectHeader(cmd *cobra.Command, _ []string) error {
 	var cnr cid.ID
 	var obj oid.ID
 
-	objAddr, err := readObjectAddress(cmd, &cnr, &obj)
+	_, err := readObjectAddress(cmd, &cnr, &obj)
 	if err != nil {
 		return err
 	}
-	mainOnly, _ := cmd.Flags().GetBool("main-only")
+
 	pk, err := key.GetOrGenerate(cmd)
 	if err != nil {
 		return err
@@ -66,9 +67,7 @@ func getObjectHeader(cmd *cobra.Command, _ []string) error {
 	}
 	defer cli.Close()
 
-	var prm internalclient.HeadObjectPrm
-	prm.SetClient(cli)
-	prm.SetPrivateKey(*pk)
+	var prm client.PrmObjectHead
 	err = Prepare(cmd, &prm)
 	if err != nil {
 		return err
@@ -79,21 +78,21 @@ func getObjectHeader(cmd *cobra.Command, _ []string) error {
 	}
 
 	raw, _ := cmd.Flags().GetBool(rawFlag)
-	prm.SetRawFlag(raw)
-	prm.SetAddress(objAddr)
-	prm.SetMainOnlyFlag(mainOnly)
+	if raw {
+		prm.MarkRaw()
+	}
 
-	res, err := internalclient.HeadObject(ctx, prm)
+	hdr, err := cli.ObjectHead(ctx, cnr, obj, user.NewAutoIDSigner(*pk), prm)
 	if err != nil {
 		if ok, err := printSplitInfoErr(cmd, err); ok {
 			return err
 		}
 		if err != nil {
-			return fmt.Errorf("rpc error: %w", err)
+			return fmt.Errorf("rpc error: read object header via client: %w", err)
 		}
 	}
 
-	err = saveAndPrintHeader(cmd, res.Header(), cmd.Flag(fileFlag).Value.String())
+	err = saveAndPrintHeader(cmd, hdr, cmd.Flag(fileFlag).Value.String())
 	if err != nil {
 		return err
 	}