+++ CHANGELOG.md
@@ -13,11 +13,13 @@ Changelog for NeoFS Node
 - Tombstone objects are removed faster after expiration now (#3511)
 
 ### Removed
+- `neofs-cli object head --main-only` no-op flag (#3509)
 
 ### Updated
 - NeoFS SDK to RC14 version fixing object inaccessibility problem (#3510)
 
 ### Updating from v0.48.1
+Remove `--main-only` flag from `neofs-cli object head` commands.
 
 ## [0.48.1] - 2025-07-31